+++ src/MudBlazor.UnitTests/Extensions/InputFileExtensions.cs
@@ -2,6 +2,7 @@
 // MudBlazor licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 using System;
+using System.Threading.Tasks;
 using Bunit;
 using Microsoft.AspNetCore.Components.Forms;
 
@@ -9,16 +10,11 @@ namespace MudBlazor.UnitTests;
 
 public static class InputFileExtensions
 {
-    public static void ClearFiles(this IRenderedComponent<InputFile> inputFileComponent)
+    public static async Task ClearFiles(this IRenderedComponent<InputFile> inputFileComponent)
     {
-        if (inputFileComponent == null)
-            throw new ArgumentNullException(nameof(inputFileComponent));
+        ArgumentNullException.ThrowIfNull(inputFileComponent);
 
         var args = new InputFileChangeEventArgs(Array.Empty<IBrowserFile>());
-        var uploadTask = inputFileComponent.InvokeAsync(() => inputFileComponent.Instance.OnChange.InvokeAsync(args));
-        if (!uploadTask.IsCompleted)
-        {
-            uploadTask.GetAwaiter().GetResult();
-        }
+        await inputFileComponent.InvokeAsync(() => inputFileComponent.Instance.OnChange.InvokeAsync(args));
     }
 }
